+++ backend/package.json
@@ -18,6 +18,6 @@
     "express-list-endpoints": "^7.1.0",
     "mongodb": "^6.6.2",
     "mongoose": "^8.4.0",
-    "nodemon": "^3.0.1"
+    "nodemon": "^3.1.1"
   }
 }
