+++ backend/server.js
@@ -1,28 +1,7 @@
-import express from "express";
-import cors from "cors";
-import mongoose from "mongoose";
+import app from "./app.js"
 
-const mongoUrl = process.env.MONGO_URL || "mongodb://localhost/OMC";
-mongoose.connect(mongoUrl);
-mongoose.Promise = Promise;
-
-// Defines the port the app will run on. Defaults to 8080, but can be overridden
-// when starting the server. Example command to overwrite PORT env variable value:
-// PORT=9000 npm start
 const port = process.env.PORT || 8080;
-const app = express();
 
-// Add middlewares to enable cors and json body parsing
-app.use(cors());
-app.use(express.json());
-
-// Start defining your routes here
-// http://localhost:8080/
-app.get("/", (req, res) => {
-  res.send("Hello Technigo!");
-});
-
-// Start the server
 app.listen(port, () => {
   console.log(`Server running on http://localhost:${port}`);
 });
